+++ packages/faceted-search/CHANGELOG.md
@@ -26,6 +26,24 @@ Types of changes
 
 ## [unreleased]
 
+## [0.2.5]
+
+### Breaking
+
+- [breaking](https://github.com/Talend/ui/pull/2599): Change equal key operator to equals
+
+### Added
+
+- [feat](https://github.com/Talend/ui/pull/2590/): Add label in add facet button
+
+### Changed
+
+- [Changed](https://github.com/Talend/ui/pull/2591): Remove Header in AddFacetPopover
+
+### Fixed
+
+- [chore](https://github.com/Talend/ui/pull/2600): change some field, snake case to camel case
+
 ## [0.2.4]
 
 ### Fixed